style: Now for write styles use scss syntax instead of sass syntax.
describe: Due to the fact that it was not possible to find an actively
supported linter for the sass syntax the project will continue to use
the scss syntax.
Other changes: All colors now define from hsla standart. It was done for
unify colors define in project.
SVG images for VK and YT is fixed.